+++ src/settings.ts
@@ -143,7 +143,7 @@ async function openSettings() {
 async function getVersion(endpoint: StashEndpoint, element: HTMLElement) {
     await request(endpoint, "version{version}", false, data => {
         element.innerHTML += `<span class="version"> (${data.version})</span>`
-    }, () => {
-        element.innerHTML += `<span class="version"> (no connection)</span>`
+    }, (message) => {
+        element.innerHTML += `<span class="version"> (${message?.trim() ?? "no connection"})</span>`
     })
 }